requirements.txt
langchain==0.3.27
langchain-community
sentence-transformers
faiss-cpu
transformers
torch
streamlit
tqdm
